+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [v1.19.5](https://github.com/contentstack/contentstack-management-javascript/tree/v1.19.5) (2025-03-17)
+ - Fix
+    - Added AuditLog in the stack class
+    - Fixed the Unit Test cases and added sanity test case for audit log
+
 ## [v1.19.4](https://github.com/contentstack/contentstack-management-javascript/tree/v1.19.4) (2025-03-10)
  - Fix
     - added fix for variants import